pyproject.toml
[tool.poetry]
name = "semester6"
version = "0.1.0"
authors = ["Nikita Shcherbak <nktshch@ya.ru>"]
readme = "README.md"
description = "The hypermodern Python project"
homepage = "https://github.com/nktshch/semester6"
keywords = ["hypermodern", "poetry"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
click = "^8.1"
requests = "^2.31.0"
matplotlib = "^3.8.4"
scikit-image = "^0.23.2"
desert = "^2022.9.22"
marshmallow = "^3.21.2"

[tool.poetry.scripts]
semester6 = "semester6.shcherbak_hw2_GHT:main"

[tool.poetry.group.dev.dependencies]
coverage = {extras = ["toml"], version = "^7.5.1"}
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
black = "^24.4.2"
flake8 = "^7.0.0"
flake8-bandit = "^4.1.1"
flake8-black = "^0.3.6"
flake8-bugbear = "^24.4.26"
flake8-import-order = "^0.18.2"
safety = "^3.2.0"
mypy = "^1.10.0"
pytype = "^2024.4.11"
typeguard = "^4.2.1"
flake8-annotations = "^3.1.0"
flake8-docstrings = "^1.7.0"
darglint = "^1.8.1"
sphinx = "^7.3.7"
sphinx-autodoc-typehints = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["semester6"]

[tool.coverage.report]
show_missing = true
